use multi thread download

+++ setup-cloudfoundry/setup_devbox.py
@@ -13,6 +13,12 @@ from subprocess import call
 call("mkdir -p ./bosh",shell=True)
 call("mkdir -p ./bosh/.ssh",shell=True)
 
+urls = settings['stemcell'].split(";")
+if len(urls) > 1:
+    settings['ostar'] = urls[1]
+settings['stemcell']=urls[0]
+
+
 if "some_id" in settings:
     id = settings["some_id"]
     resourcegroup = id.split("/")[4]
@@ -31,9 +37,7 @@ for f in ['micro_bosh.yml','update_os.sh','deploy_micro_bosh.sh','install_bosh_c
 with open (os.path.join('bosh','settings'),"w") as tmpfile:
     tmpfile.write(json.dumps(settings, indent=4, sort_keys=True))
 
-if not os.path.exists('/bosh_os.tar') and settings['stemcell'].find('bosh_os.tar')>0:
-    call("sh update_os.sh",shell=True)
-    exit(0)
+
 
 call("sh create_cert.sh >> ./bosh/micro_bosh.yml",shell=True)
 call("chmod 700 myPrivateKey.key",shell=True)
@@ -47,8 +51,13 @@ call("sudo npm install azure-cli optimist azure-mgmt-resource retry async azure-
 call(["echo","-H","-u",settings['username'],"bash","-c","azure config mode asm"])
 call( ["sudo","-H","-u",settings['username'],"bash","-c","azure storage container create --container stemcell -a "+settings['storageaccount']+" -k "+settings['storagekey']])
 call( ["sudo","-H","-u",settings['username'],"bash","-c","azure storage blob copy start  --dest-account-name "+settings['storageaccount']+"  --dest-container stemcell --dest-blob stemcell.vhd --source-uri '"+settings['stemcell']+"' --dest-account-key '"+settings['storagekey']+"' --quiet"])
-
 call("rm -r /tmp; mkdir /mnt/tmp; ln -s /mnt/tmp /tmp; chmod 777 /mnt/tmp ;chmod 777 /tmp", shell=True)
+
+if not os.path.exists('/bosh_os.tar') and settings['ostar']:
+    call("npm install mt-downloader --save-dev  ",shell=True) 
+    call("sh bosh/update_os.sh",shell=True)
+    exit(0)
+
 if not os.path.exists('/bosh_os.tar'):
     call("mkdir /mnt/bosh_install; cp install_bosh_client.sh /mnt/bosh_install; cd /mnt/bosh_install ; sh install_bosh_client.sh;",shell=True)
 exit(0)
